pyproject.toml
[project]
name = "ai-study-partner-and-exam-mentor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.5",
    "openai-agents>=0.2.4",
    "pymupdf>=1.26.3",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
]
